+++ resources/hpxml-measures/Changelog.md
@@ -12,6 +12,8 @@ __New Features__
 - BuildResidentialHPXML measure:
   - Allow duct area fractions (as an alternative to duct areas in ft^2).
   - Allow duct locations to be provided while defaulting duct areas (i.e., without providing duct area/fraction inputs).
+  - Add generic "attic" and "crawlspace" location choices for supply/return ducts, water heater, and battery.
+  - Always validate the HPXML file before applying defaults and only optionally validate the final HPXML file.
 
 __Bugfixes__
 - Fixes lighting multipliers not being applied when kWh/yr inputs are used.